requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopage==0.5.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cliff==4.1.0
cmaes==0.9.1
cmd2==2.4.2
colorlog==6.7.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
greenlet==2.0.1
huggingface-hub==0.11.1
hydra-colorlog==1.2.0
hydra-core==1.3.1
hydra-optuna-sweeper==1.2.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
lightning-utilities==0.5.0
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
optuna==2.10.1
packaging==23.0
pandas==1.5.2
parso==0.8.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prettytable==3.6.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrootutils==1.0.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.8.6
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==13.0.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
stevedore==4.1.1
tensorboardX==2.5.1
timm==0.6.12
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
yarl==1.8.2
zipp==3.11.0
